Use .NET Core 3.1
These changes the project to be a .NET Core 3.1 Console App.

If this project is opened in Visual Studio Code then you can
run either of the programs with a debugger from the GUI by select
the launch configuration.

The switch was made so that parsing features could take advantage
of nuget packages such as System.Text.Json which are implemented
.Net Core 3.1.